--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -1,5 +1,8 @@
-module github.com/fixme_my_friend/hw10_program_optimization
+module github.com/galogen13/otus_hw/hw10_program_optimization
 
 go 1.14
 
-require github.com/stretchr/testify v1.5.1
+require (
+	github.com/mailru/easyjson v0.7.1
+	github.com/stretchr/testify v1.5.1
+)