--- a/hw10_program_optimization/stats_easyjson.go
+++ b/hw10_program_optimization/stats_easyjson.go
@@ -36,20 +36,8 @@ func easyjsonE3ab7953DecodeGithubComGalogen13OtusHwHw10ProgramOptimization(in *j
 			continue
 		}
 		switch key {
-		case "ID":
-			out.ID = int(in.Int())
-		case "Name":
-			out.Name = string(in.String())
-		case "Username":
-			out.Username = string(in.String())
 		case "Email":
 			out.Email = string(in.String())
-		case "Phone":
-			out.Phone = string(in.String())
-		case "Password":
-			out.Password = string(in.String())
-		case "Address":
-			out.Address = string(in.String())
 		default:
 			in.SkipRecursive()
 		}
@@ -65,39 +53,9 @@ func easyjsonE3ab7953EncodeGithubComGalogen13OtusHwHw10ProgramOptimization(out *
 	first := true
 	_ = first
 	{
-		const prefix string = ",\"ID\":"
-		out.RawString(prefix[1:])
-		out.Int(int(in.ID))
-	}
-	{
-		const prefix string = ",\"Name\":"
-		out.RawString(prefix)
-		out.String(string(in.Name))
-	}
-	{
-		const prefix string = ",\"Username\":"
-		out.RawString(prefix)
-		out.String(string(in.Username))
-	}
-	{
 		const prefix string = ",\"Email\":"
-		out.RawString(prefix)
+		out.RawString(prefix[1:])
 		out.String(string(in.Email))
-	}
-	{
-		const prefix string = ",\"Phone\":"
-		out.RawString(prefix)
-		out.String(string(in.Phone))
-	}
-	{
-		const prefix string = ",\"Password\":"
-		out.RawString(prefix)
-		out.String(string(in.Password))
-	}
-	{
-		const prefix string = ",\"Address\":"
-		out.RawString(prefix)
-		out.String(string(in.Address))
 	}
 	out.RawByte('}')
 }