--- a/hw10_program_optimization/stats_test.go
+++ b/hw10_program_optimization/stats_test.go
@@ -3,7 +3,9 @@
 package hw10_program_optimization //nolint:golint,stylecheck
 
 import (
+	"archive/zip"
 	"bytes"
+	"log"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -15,6 +17,18 @@ func TestGetDomainStat(t *testing.T) {
 {"Id":3,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
 {"Id":4,"Name":"Gregory Reid","Username":"tButler","Email":"5Moore@Teklist.net","Phone":"520-04-16","Password":"r639qLNu","Address":"Sunfield Park 20"}
 {"Id":5,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@Linktype.com","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}`
+
+	uncorrectData := `Bad data
+Bad data
+Bad data
+Bad data
+Bad data
+Bad data
+Bad data
+Bad data
+Bad data
+Bad data
+Bad data`
 
 	t.Run("find 'com'", func(t *testing.T) {
 		result, err := GetDomainStat(bytes.NewBufferString(data), "com")
@@ -36,4 +50,44 @@ func TestGetDomainStat(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, DomainStat{}, result)
 	})
+
+	t.Run("empty data", func(t *testing.T) {
+		result, err := GetDomainStat(bytes.NewBufferString(""), "com")
+		require.NoError(t, err)
+		require.Equal(t, DomainStat{}, result)
+	})
+
+	t.Run("empty domain", func(t *testing.T) {
+		result, err := GetDomainStat(bytes.NewBufferString(data), "")
+		require.Equal(t, ErrEmptyDomain, err)
+		require.Nil(t, result)
+	})
+
+	t.Run("no JSON data", func(t *testing.T) {
+		result, err := GetDomainStat(bytes.NewBufferString(uncorrectData), "com")
+		require.NotNil(t, err)
+		require.Nil(t, result)
+	})
+}
+
+// BenchmarkGetDomainStat .
+func BenchmarkGetDomainStat(b *testing.B) {
+
+	r, err := zip.OpenReader("testdata/users.dat.zip")
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer r.Close()
+
+	data, err := r.File[0].Open()
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		GetDomainStat(data, "com")
+
+	}
+
 }