benchmark                    old ns/op      new ns/op     delta
BenchmarkGetDomainStat-2     1128252102     7642          -100.00%
BenchmarkGetDomainStat-2     1195459880     7543          -100.00%
BenchmarkGetDomainStat-2     1186286999     7683          -100.00%
BenchmarkGetDomainStat-2     1173494015     10774         -100.00%
BenchmarkGetDomainStat-2     1152922585     7262          -100.00%
BenchmarkGetDomainStat-2     1148660640     10241         -100.00%
BenchmarkGetDomainStat-2     1179490490     8579          -100.00%
BenchmarkGetDomainStat-2     1166699318     8456          -100.00%
BenchmarkGetDomainStat-2     1158656781     7082          -100.00%
BenchmarkGetDomainStat-2     1161696741     7862          -100.00%

benchmark                    old allocs     new allocs     delta
BenchmarkGetDomainStat-2     3045558        14             -100.00%
BenchmarkGetDomainStat-2     3045553        14             -100.00%
BenchmarkGetDomainStat-2     3045545        14             -100.00%
BenchmarkGetDomainStat-2     3045552        14             -100.00%
BenchmarkGetDomainStat-2     3045553        14             -100.00%
BenchmarkGetDomainStat-2     3045556        15             -100.00%
BenchmarkGetDomainStat-2     3045556        15             -100.00%
BenchmarkGetDomainStat-2     3045553        14             -100.00%
BenchmarkGetDomainStat-2     3045549        14             -100.00%
BenchmarkGetDomainStat-2     3045550        14             -100.00%

benchmark                    old bytes     new bytes     delta
BenchmarkGetDomainStat-2     285160512     4988          -100.00%
BenchmarkGetDomainStat-2     285086576     5033          -100.00%
BenchmarkGetDomainStat-2     285012112     4987          -100.00%
BenchmarkGetDomainStat-2     285086400     4999          -100.00%
BenchmarkGetDomainStat-2     285159288     5021          -100.00%
BenchmarkGetDomainStat-2     285123568     5083          -100.00%
BenchmarkGetDomainStat-2     285123296     5124          -100.00%
BenchmarkGetDomainStat-2     285086256     5051          -100.00%
BenchmarkGetDomainStat-2     285048984     4996          -100.00%
BenchmarkGetDomainStat-2     285085768     5041          -100.00%

## Домашнее задание №10 «Оптимизация программы»

### Критерии оценки
- [x] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [x] Добавлены юнит-тесты - до 3 баллов
- [x] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов