@galogen13 2020-05-31
**NEW BENCHCMP**
benchmark                    old ns/op      new ns/op     delta
BenchmarkGetDomainStat-2     1128252102     6917          -100.00%
BenchmarkGetDomainStat-2     1195459880     14933         -100.00%
BenchmarkGetDomainStat-2     1186286999     15170         -100.00%
BenchmarkGetDomainStat-2     1173494015     8368          -100.00%
BenchmarkGetDomainStat-2     1152922585     10954         -100.00%
BenchmarkGetDomainStat-2     1148660640     9721          -100.00%
BenchmarkGetDomainStat-2     1179490490     7468          -100.00%
BenchmarkGetDomainStat-2     1166699318     7634          -100.00%
BenchmarkGetDomainStat-2     1158656781     7377          -100.00%
BenchmarkGetDomainStat-2     1161696741     7714          -100.00%

benchmark                    old allocs     new allocs     delta
BenchmarkGetDomainStat-2     3045558        10             -100.00%
BenchmarkGetDomainStat-2     3045553        10             -100.00%
BenchmarkGetDomainStat-2     3045545        13             -100.00%
BenchmarkGetDomainStat-2     3045552        11             -100.00%
BenchmarkGetDomainStat-2     3045553        12             -100.00%
BenchmarkGetDomainStat-2     3045556        11             -100.00%
BenchmarkGetDomainStat-2     3045556        11             -100.00%
BenchmarkGetDomainStat-2     3045553        11             -100.00%
BenchmarkGetDomainStat-2     3045549        11             -100.00%
BenchmarkGetDomainStat-2     3045550        11             -100.00%

benchmark                    old bytes     new bytes     delta
BenchmarkGetDomainStat-2     285160512     4733          -100.00%
BenchmarkGetDomainStat-2     285086576     4737          -100.00%
BenchmarkGetDomainStat-2     285012112     4934          -100.00%
BenchmarkGetDomainStat-2     285086400     4767          -100.00%
BenchmarkGetDomainStat-2     285159288     4832          -100.00%
BenchmarkGetDomainStat-2     285123568     4752          -100.00%
BenchmarkGetDomainStat-2     285123296     4755          -100.00%
BenchmarkGetDomainStat-2     285086256     4763          -100.00%
BenchmarkGetDomainStat-2     285048984     4760          -100.00%
BenchmarkGetDomainStat-2     285085768     4745          -100.00%