@galogen13 2020-06-03
benchmark                    old ns/op      new ns/op     delta
BenchmarkGetDomainStat-2     1128252102     2349          -100.00%
BenchmarkGetDomainStat-2     1195459880     2392          -100.00%
BenchmarkGetDomainStat-2     1186286999     2336          -100.00%
BenchmarkGetDomainStat-2     1173494015     2343          -100.00%
BenchmarkGetDomainStat-2     1152922585     2991          -100.00%
BenchmarkGetDomainStat-2     1148660640     5864          -100.00%
BenchmarkGetDomainStat-2     1179490490     6862          -100.00%
BenchmarkGetDomainStat-2     1166699318     1898          -100.00%
BenchmarkGetDomainStat-2     1158656781     2054          -100.00%
BenchmarkGetDomainStat-2     1161696741     2448          -100.00%

benchmark                    old allocs     new allocs     delta
BenchmarkGetDomainStat-2     3045558        2              -100.00%
BenchmarkGetDomainStat-2     3045553        2              -100.00%
BenchmarkGetDomainStat-2     3045545        2              -100.00%
BenchmarkGetDomainStat-2     3045552        2              -100.00%
BenchmarkGetDomainStat-2     3045553        2              -100.00%
BenchmarkGetDomainStat-2     3045556        3              -100.00%
BenchmarkGetDomainStat-2     3045556        3              -100.00%
BenchmarkGetDomainStat-2     3045553        2              -100.00%
BenchmarkGetDomainStat-2     3045549        2              -100.00%
BenchmarkGetDomainStat-2     3045550        2              -100.00%

benchmark                    old bytes     new bytes     delta
BenchmarkGetDomainStat-2     285160512     4193          -100.00%
BenchmarkGetDomainStat-2     285086576     4187          -100.00%
BenchmarkGetDomainStat-2     285012112     4192          -100.00%
BenchmarkGetDomainStat-2     285086400     4193          -100.00%
BenchmarkGetDomainStat-2     285159288     4202          -100.00%
BenchmarkGetDomainStat-2     285123568     4243          -100.00%
BenchmarkGetDomainStat-2     285123296     4280          -100.00%
BenchmarkGetDomainStat-2     285086256     4184          -100.00%
BenchmarkGetDomainStat-2     285048984     4182          -100.00%
BenchmarkGetDomainStat-2     285085768     4194          -100.00%